Move the testing code from Teacher's example(where it shouldn't be) to JAVA/JUnit test case, and also make jsp work again.

+++ JAVA/springboot/src/main/java/com/JoeMob/springboot/controller/bean/StudentBean.java
@@ -11,7 +11,7 @@ import java.util.Date;
 @Data
 public class StudentBean {
     private Long id;
-    private Integer studentId;
+    private Long studentId;
     private String studentName;
     private Integer age;
     private String gender;
